Move App out of UI; Fix imports
